0311: react project clear up, hook intro to useState
