+++ index.css
@@ -27,7 +27,7 @@ body{
 }
 
 .app-body {
-    padding: 10px 130px;
+    padding: 10px 30px;
 }
 
 .restaraunt-container {
@@ -35,8 +35,9 @@ body{
     flex-wrap: wrap;
 }
 
-.search {
+.filter {
     margin: 20px 0;
+    padding: 10px;
 }
 
 .restaurant-card {
