+++ index.html
@@ -12,6 +12,6 @@
             <h3>Not Rendered</h3>
         </div>
         <!-- <script src="./script.js" type="module"></script> -->
-        <script src="./App.js" type="module"></script>
+        <script src="./src/App.js" type="module"></script>
     </body>
 </html>
