+++ notes.md
@@ -1,0 +1,16 @@
+```js
+/*
+Header
+    - Logo 
+    - nav menu
+body
+    - Search
+    - Restaurant Container
+        - Restaurant cards
+Footer
+    - Copyright 
+    - links 
+    - Address
+    - Contacts
+*/
+```
