+++ src/utils/mockData.js
@@ -1,44 +1,4 @@
-import React from "react";
-import ReactDOM from "react-dom/client";
-import logo from "./public/asset/food_app_logo.png";
-import resImg from "./public/asset/Biryani_2.avif";
-
-/*
-Header
-    - Logo 
-    - nav menu
-body
-    - Search
-    - Restaurant Container
-        - Restaurant cards
-Footer
-    - Copyright 
-    - links 
-    - Address
-    - Contacts
-*/
-
-const HeaderComponent = () => (
-    <div className="header">
-        <div className="logo-container">
-            <img className="logo" src={logo} />
-        </div>
-        <div className="nav-items">
-            <ul className="nav">
-                <li>Home</li>
-                <li>About Us</li>
-                <li>Contact Us</li>
-                <li>Cart</li>
-            </ul>
-        </div>
-    </div>
-);
-
-// const cardStyle = {
-//     backgroundColor: "#f0f0f0",
-// };
-
-const res_data = [
+export const MOCK_DATA = [
     {
         info: {
             id: "690702",
@@ -689,67 +649,3 @@ const res_data = [
         widgetId: "collectionV5RestaurantListWidget_SimRestoRelevance_food_seo",
     },
 ];
-
-const RestaurantCardComponent = (props) => {
-    const { resData } = props;
-    return (
-        <div
-            className="restaurant-card"
-            style={{
-                backgroundColor: "#f0f0f0",
-                padding: "5px",
-            }}
-        >
-            <img
-                src={
-                    "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/" +
-                    resData.info.cloudinaryImageId
-                }
-                style={{ width: "100%" }}
-            />
-            <h3>{resData.info.name}</h3>
-            <h5>{resData.info.cuisines.join(", ")}</h5>
-            <div>{resData.info.avgRating} star</div>
-            <div>{resData.info.sla.deliveryTime} minutes</div>
-        </div>
-    );
-};
-
-const BodyComponent = () => (
-    <div className="app-body">
-        <div className="search">
-            <input type="text" />
-            <button type="button" />
-        </div>
-        <div className="restaraunt-container">
-            {res_data.map((data) => {
-                return (
-                    <RestaurantCardComponent
-                        key={data.info.id}
-                        resData={data}
-                    />
-                );
-            })}
-            {/* <RestaurantCardComponent resData={res_data[0]} />
-            <RestaurantCardComponent resData={res_data[1]} />
-            <RestaurantCardComponent resData={res_data[2]} />
-            <RestaurantCardComponent resData={res_data[3]} /> */}
-            {/* <RestaurantCardComponent resname="Aarya Bhavan" /> */}
-        </div>
-    </div>
-);
-
-const AppLayout = () => {
-    return (
-        <div id="app">
-            <HeaderComponent />
-            <BodyComponent />
-            {/* Header
-            body
-            footer */}
-        </div>
-    );
-};
-
-const app = ReactDOM.createRoot(document.getElementById("root"));
-app.render(<AppLayout />);
